backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
psycopg2-binary==2.9.11
pytesseract==0.3.13
pdf2image==1.17.0
PyPDF2==3.0.1
python-magic-bin==0.4.14
